Fix constant evaluation logic for String.length.
Previously, we used static resolution to determine whether
String.length was being referred to, causing incorrect results in the
presence of bad type annotations.  Also, we were missing a null check,
causing constant evaluation to crash if an attempt was made to take
the length of a constant that couldn't be evaluated.

BUG=dartbug.com/23383
R=scheglov@google.com

Review URL: https://codereview.chromium.org//1124433007

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@45524 260f80e4-7a28-3924-810f-c04153c831b5