Ensure that nulls aren't added to constant eval dependency graph.
In r45541, I accidentally left out a null check causing nulls to get
inserted into the constant evaluation dependency graph.  This caused a
log message to be printed when running the command-line analyzer, but
no test failures.

I've inserted the proper null check and added an assert statement so
that if the bug recurs it will be caught during testing.

BUG=dartbug.com/23407
R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//1133623002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@45594 260f80e4-7a28-3924-810f-c04153c831b5