Avoid using multitest.
This test did not work as a multitest, because the
deferred_redirecting_factory_lib1.dart imports the main file - it will not
find it after the splitting and renaming.

Luckily it is not strictly required for the test (though somewhat cleaner)
to have a multitest.

BUG=

Review URL: https://codereview.chromium.org//1127023003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@45643 260f80e4-7a28-3924-810f-c04153c831b5