Add a "librarySource" field to CompilationUnitElementImpl.
This fixes a dependency problem in ComputeConstantDependenciesTask.
Prevously it used element.library.source to get access to the library
source, but this created an implicit dependency on LIBRARY_ELEMENT1
(since element.library is null if the library element hasn't been
built yet).  There was no way to make that dependency explicit since
previous to this change, the CompilationUnitElement did not store
enough information to determine what library it was part of until
after the library element had been created.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//1125123006

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@45674 260f80e4-7a28-3924-810f-c04153c831b5