In the new task model, use explicit dependencies on the type provider.
Previously, some of the tasks got the type provider from the context,
which meant we were at risk of invoking the task manager reentrantly
(which might have broken some task manager assumptions and potentially
degraded responsiveness).  With this CL, all tasks that need the type
provider get it using an explicit input.

Assertions are added so that if we accidentally introduce future code
paths that invoke the task manager in a reentrant fashion, we will see
the problem during unit tests.

R=brianwilkerson@google.com, scheglov@google.com

Review URL: https://codereview.chromium.org//1124003007

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@45794 260f80e4-7a28-3924-810f-c04153c831b5